[MovieList] Make buildBeginTimeSortKey() return a 3 item tuple
buildGroupwiseSortkey() mixes the return values of buildBeginTimeSortKey() and buildAlphaNumericSortKey() so both need to create a tuple of the same length, and each item in the tuple should be the same data type i.e. (int, str, int/float). If these don't correspond properly it will lead to crashes.

Thx Huevos